Moved login.scss.
