+++ src/js/verify/verify-entry.jsx
@@ -1,10 +1,7 @@
-import '../../platform/polyfills';
-import '../../sass/login.scss';
-
 import React from 'react';
 
+import '../../platform/polyfills';
 import startApp from '../../platform/startup';
-
 import VerifyApp from './containers/VerifyApp';
 
 startApp({ component: <VerifyApp/> });
