+++ src/platform/site-wide/login/login-entry.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 
-import '../../../sass/login.scss';
+import './sass/login.scss';
 import startReactApp from '../../startup/react';
 import Main from './containers/Main';
